shift some more code around
